+++ src/main/java/blackjack/domain/gambler/Dealer.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.gambler;
 
 import blackjack.domain.card.Card;
 import java.util.List;
