+++ src/main/java/blackjack/domain/gambler/Player.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.gambler;
 
 import java.util.Objects;
